Update: PickUp Gun
It works now, you can pick up a gun, but must clearly define the script of each slot so it can grab the correct data of location it will be place at, which is the player hands (for now), there is two guns to use atm for testing, one white, one with some color. (future, there will be changes to the code to get it working with any other code or how we want it to be at, location and etc and bullets.