Update: New Gun Model (any bug encounter fixed)
New Gun Model was added, there was some bugs when added but that was fixed, still an bug issue when throwing a gun, it wont be replace but 2nd slot gun so you must hit tab to bring it out otherwise it works. and there is another bug which cause one gun on ground still connected, gotta fix that later.