Update: Rename Game Scene and Add MenuScene
Fix a problem with the switching weapon and pick and drop weapon, the only thing that had to be added was adding isReloading = false in OnDisable function which fix all issues so the gun is reloading even at 0.
I've rename the old Scene which was called SampleScene to GameScene and added a new Scene called Menu which has the UI menu of play, option and exit, all working.