Update Fix: Fixed Upgrade Weapon aka PowerUp
Fixed PowerUp where I mistake comment the needed code that was created but was mistaken and commented out, so it was uncomment now. works wonders for any weapon that is current main weapon to get a powerup increase bullet damage aka new bullet that has new damage value compare to starting one.